moved helper functions into new file protection_coverage.py
